backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
supabase==2.10.0
python-dotenv==1.0.1
